chore: replaced primary key identifier to long

+++ src/main/java/com/NeuralN/VibesRa/controller/ReportController.java
@@ -22,7 +22,7 @@ public class ReportController {
     }
 
     @GetMapping("/{reportId}")
-    public ResponseEntity<Report> getReportById(@PathVariable int reportId) {
+    public ResponseEntity<Report> getReportById(@PathVariable Long reportId) {
         Report report = reportService.getReportById(reportId);
         if (report == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -36,7 +36,7 @@ public class ReportController {
     }
 
     @DeleteMapping("/{reportId}")
-    public ResponseEntity<Void> deleteReport(@PathVariable int reportId) {
+    public ResponseEntity<Void> deleteReport(@PathVariable Long reportId) {
         reportService.deleteReport(reportId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
