+++ src/main/java/com/NeuralN/VibesRa/repository/ImageRepository.java
@@ -3,5 +3,5 @@ package com.NeuralN.VibesRa.repository;
 import com.NeuralN.VibesRa.model.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageRepository extends JpaRepository<Image, Integer> {
+public interface ImageRepository extends JpaRepository<Image, Long> {
 }
